requirements.txt
streamlit==1.28.0
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0

# Audio processing
SpeechRecognition==3.10.0
PyAudio==0.2.11
gTTS==2.3.0

# Alternative audio backends (choose based on your system)
# pocketsphinx==5.0.0  # For offline speech recognition
# pyaudio-fork==0.2.11  # Alternative PyAudio

# AWS and ML
langchain==0.1.0
langchain-community==0.0.10